+++ prune_via_test_sample.py
@@ -1,4 +1,3 @@
-# Importing Libraries
 import argparse
 import os
 from tqdm import tqdm
@@ -8,7 +7,7 @@ import numpy as np
 import torch.nn as nn
 import torch.nn.init as init
 #from tensorboardX import SummaryWriter
-
+import ipdb
 # Custom Libraries
 from utility.data import *
 from utility.log_utils import *
@@ -32,7 +31,7 @@ def main(args):
   #  device = torch.device("cuda:1")
     reinit = True if args.initial=="reinit" else False
 
-    # Data Loader 
+    # Data Loader
     if args.dataset.lower() == "cifar10":
         size = (32, 32)
         dataset = Cifar10(args.batch_size, args.threads, size, args.augmentation)
@@ -44,7 +43,7 @@ def main(args):
     elif args.dataset.lower() == "cifar100":
         size = (32, 32)
         dataset = Cifar100(args.batch_size, args.threads, size, args.augmentation)
-        from archs.cifar100 import AlexNet, fc1, LeNet5, vgg, resnet 
+        from archs.cifar100 import AlexNet, fc1, LeNet5, vgg, resnet
     else:
         size = (28, 28)
         dataset = MNIST(args.batch_size, args.threads, size, args.augmentation)
@@ -58,20 +57,20 @@ def main(args):
     elif args.arch_type.lower() == "alexnet":
         model = AlexNet.AlexNet().to(device)
     elif args.arch_type.lower() == "vgg16":
-        model = vgg.VGG("VGG16").to(device)  
+        model = vgg.VGG("VGG16").to(device)
     elif args.arch_type.lower() == "resnet18":
-        model = resnet.resnet18().to(device)   
+        model = resnet.resnet18().to(device)
     elif args.arch_type.lower() == "resnet50":
-        model = resnet.resnet50().to(device)   
+        model = resnet.resnet50().to(device)
     elif args.arch_type.lower() == "fcs":
-        model = fcs.fcs().to(device)    
+        model = fcs.fcs().to(device)
     else:
         print("\nWrong Model choice\n")
         exit()
 
     # Making Initial Mask
     mask = make_mask(model)
-
+#    ipdb.set_trace()
     # Initial Model
     model_init = copy.deepcopy(model)
     # Copying and Saving Initial State
@@ -95,51 +94,61 @@ def main(args):
     prior_sigma = args.prior
     kl=args.kl
     noise_step = args.noise_step
-    
+
     p_old = None
     model_schedule = None
     for _ite in range(args.start_iter, ITERATION):
-        
+
+        #for batch_idx, (data, target) in enumerate(dataset.test1):
+        #    data, target = data.to(device), target.to(device)
+        #    print(np.histogram(target.detach().cpu().int().numpy(),bins=10))
         if not _ite == 0:
-            ## prune here 
+            if _ite ==1:
+                model_unprune = copy.deepcopy(model)
+            ## prune here
             if args.prune_type=="noise":
                # numb_w =0
                # for i in range(len(mask)):
-                
-                mask, p_new, p_schedule = prune_by_noise(model, mask, args.prune_percent, dataset.train,criterion,noise_type,
+               # ipdb.set_trace()
+               if _ite < 10:
+                    mask = prune_by_percentile(model, mask, args.prune_percent)
+               else:
+                    mask, p_new, p_schedule = prune_by_noise(model,model_unprune, mask, args.prune_percent, dataset.test1,criterion,noise_type,
                         prior_sigma,kl,num_steps=noise_step,lr=args.lr_p, p_init=p_old, reduce_op=args.reduce_kl)
-                all_masks = torch.cat([m.view(-1) for m in mask])
-                print(torch.norm(all_masks*p_new,1)/torch.norm(all_masks*p_new,2))
-                print(p_new[all_masks.bool()])
-                if args.initial_p == "last":
-                    p_old = p_new.detach().clone()
-                elif args.initial_p == "schedule":
-                    p_old = p_schedule.detach().clone()
-                else:
-                    _,p_old,_ ,_= initialization(model,mask,args.prior,args.noise_type)
+                    all_masks = torch.cat([m.view(-1) for m in mask])
+                    print(torch.norm(all_masks*p_new,1)/torch.norm(all_masks*p_new,2))
+                    print(p_new[all_masks.bool()])
+                    if args.initial_p == "last":
+                        p_old = p_new.detach().clone()
+                    elif args.initial_p == "schedule":
+                        p_old = p_schedule.detach().clone()
+                    else:
+                        _,p_old,_ ,_= initialization(model,mask,args.prior,args.noise_type)
 
             elif args.prune_type=="noise_pac":
                 # reweight weight
                 with torch.no_grad():
                     for i, (param1, param2) in enumerate(zip(model.parameters(), model_init.parameters())):
-                        param2.data = (torch.norm(param1.data * mask[i]) / 
-                                                    (torch.norm(param2.data * mask[i])+1e-6) * 
+                        param2.data = (torch.norm(param1.data * mask[i]) /
+                                                    (torch.norm(param2.data * mask[i])+1e-6) *
                                                     param2.data * mask[i])
 
                 mask, p_new, p_schedule  = prune_by_noise_trainable_prior(model, model_init, mask, args.prune_percent, dataset.train,criterion,
-                    num_steps=noise_step,lr=args.lr_p, p_init=p_old)   
+                    num_steps=noise_step,lr=args.lr_p, p_init=p_old)
                 if args.initial_p == "last":
                     p_old = p_new.detach().clone()
                 elif args.initial_p == "schedule":
                     p_old = p_schedule.detach().clone()
-                                 
-            elif args.prune_type=="lt":    
+
+            elif args.prune_type=="lt":
                 mask = prune_by_percentile(model, mask, args.prune_percent)
+            elif args.prune_type=="lt_fine":
+                mask = prune_by_percentile_fine_tune(model,model_unprune,dataset.test1, args.lr, args.weight_decay,criterion, args.enditer, mask, args.prune_percent)
             elif args.prune_type=="random":
                 mask = prune_by_random(model, mask, args.prune_percent)
 
 
-            ## initialize here 
+            ## initialize here
             if args.initial=="reinit":
                 reset_all_weights(model)
                 step = 0
@@ -156,7 +165,7 @@ def main(args):
                     original_initialization(model, mask, copy.deepcopy(model))
                 model_schedule = None
             elif args.initial=="last":
-                original_initialization(model, mask, copy.deepcopy(model)) ## does not alter the model, only masks it 
+                original_initialization(model, mask, copy.deepcopy(model)) ## does not alter the model, only masks it
             elif args.initial=="rewind":
                 print("initialization at rewind")
                 ## load the model from the rewind folder
@@ -164,14 +173,14 @@ def main(args):
                 ## mask the model
                 original_initialization(model, mask, copy.deepcopy(model_rewind))
 
-        count_nonzero(model, mask)   
-        if args.optimizer.lower() == "sgd":         
-            optimizer = torch.optim.SGD(model.parameters(), lr=args.lr, 
+        count_nonzero(model, mask)
+        if args.optimizer.lower() == "sgd":
+            optimizer = torch.optim.SGD(model.parameters(), lr=args.lr,
                                             momentum = args.momentum,
-            								 weight_decay = args.weight_decay)
-        else:     
-            optimizer = torch.optim.Adam(model.parameters(), lr=args.lr, 
-                                         weight_decay = args.weight_decay)           
+                                                                         weight_decay = args.weight_decay)
+        else:
+            optimizer = torch.optim.Adam(model.parameters(), lr=args.lr,
+                                         weight_decay = args.weight_decay)
 
         scheduler = ReduceLROnPlateau(optimizer, mode='max', factor=0.1, patience=10)
         print(f"\n--- Pruning Level [{_ite}/{ITERATION}]: ---")
@@ -180,59 +189,65 @@ def main(args):
         comp[_ite] = comp1
         if _ite == 0:
 
+          if os.path.isfile(f"{os.getcwd()}/saves/{args.arch_type}/{args.dataset}/{args.kl}+{args.prior}/{_ite}_model_{args.prune_type}.pth.tar"):
+            model = torch.load(f"{os.getcwd()}/saves/{args.arch_type}/{args.dataset}/{args.kl}+{args.prior}/{_ite}_model_{args.prune_type}.pth.tar")
+          else:
             pbar = tqdm(range(end_iter))
-        else:
-            pbar = tqdm(range(args.enditer))
 
-        achieve_target_acc, tolerance = 0, 0
-        for iter_ in pbar:
+            achieve_target_acc, tolerance = 0, 0
+
+            for iter_ in pbar:
 
             # Frequency for Testing
-            if iter_ % args.valid_freq == 0:
-                accuracy = test(model, dataset.test, criterion)
-                train_acc = test(model, dataset.train, criterion)
+              if iter_ % args.valid_freq == 0:
+                  accuracy = test(model, dataset.test2, criterion)
+                  train_acc = test(model, dataset.train, criterion)
 
                 # Save Weights
-                if accuracy >= best_accuracy:
-                    best_accuracy = accuracy
-                    checkdir(f"{os.getcwd()}/saves/{args.arch_type}/{args.dataset}/{args.kl}+{args.prior}/")
-                    torch.save(model,f"{os.getcwd()}/saves/{args.arch_type}/{args.dataset}/{args.kl}+{args.prior}/{_ite}_model_{args.prune_type}.pth.tar")
+                  if accuracy >= best_accuracy:
+                      best_accuracy = accuracy
+                      checkdir(f"{os.getcwd()}/saves/{args.arch_type}/{args.dataset}/{args.kl}+{args.prior}/")
+                      torch.save(model,f"{os.getcwd()}/saves/{args.arch_type}/{args.dataset}/{args.kl}+{args.prior}/{_ite}_model_{args.prune_type}.pth.tar")
 
             # Training
-            if args.noise_std == 0:
-                loss = train(model, mask, dataset.train, optimizer, criterion,args.l1)
-            else:
-                loss = train_with_noise(model, mask, dataset.train, optimizer, criterion, 
+              if args.noise_std == 0:
+                   loss = train(model, mask, dataset.train, optimizer, criterion)
+              else:
+                  loss = train_with_noise(model, mask, dataset.train, optimizer, criterion,
                                             noise_std=args.noise_std, noise_type=args.inject_noise)
 
-            all_loss[iter_] = loss
-            all_accuracy[iter_] = accuracy
+              all_loss[iter_] = loss
+              all_accuracy[iter_] = accuracy
 
-            if train_acc >= 99.9:
-                achieve_target_acc += 1
-                if achieve_target_acc > 20:
-                    break
+              if train_acc >= 99.9:
+                  achieve_target_acc += 1
+                  if achieve_target_acc > 20:
+                      break
 
             # no need to keep training
-            if optimizer.param_groups[0]['lr'] < 1e-5:
-                break
-            scheduler.step(train_acc)
-            if args.initial == "schedule":
-                if optimizer.param_groups[0]['lr'] < args.lr - 1e-6 and model_schedule is None:
-                    model_schedule = copy.deepcopy(model)
+              if optimizer.param_groups[0]['lr'] < 1e-5:
+                   break
+              scheduler.step(train_acc)
+              if args.initial == "schedule":
+                  if optimizer.param_groups[0]['lr'] < args.lr - 1e-6 and model_schedule is None:
+                      model_schedule = copy.deepcopy(model)
 
-            if args.initial=="rewind" and args.rewind_iter==iter_ and _ite==0:
+              if args.initial=="rewind" and args.rewind_iter==iter_ and _ite==0:
                 ## save the model in a separaate folder named rewind
                 checkdir(f"{os.getcwd()}/saves/{args.arch_type}/{args.dataset}/rewind/{args.kl}+{args.prior}")
                 torch.save(model,f"{os.getcwd()}/saves/{args.arch_type}/{args.dataset}/rewind/{args.kl}+{args.prior}/{args.rewind_iter}_model_{args.prune_type}_{args.noise_type}.pth.tar")
                 print("here at",iter_)
-            
-            # Frequency for Printing Accuracy and Loss
-            if iter_ % args.print_freq == 0:
-                pbar.set_description(
-                    f'Train Epoch: {iter_}/{end_iter} Loss: {loss:.6f} Accuracy: {accuracy:.2f} {train_acc:.2f} % Best test Accuracy: {best_accuracy:.2f}%')       
 
-        print(f'Train Epoch: {iter_}/{end_iter} Loss: {loss:.6f} Accuracy: {accuracy:.2f} {train_acc:.2f} % Best test Accuracy: {best_accuracy:.2f}%') 
+            # Frequency for Printing Accuracy and Loss
+              if iter_ % args.print_freq == 0:
+                 pbar.set_description(
+                    f'Train Epoch: {iter_}/{end_iter} Loss: {loss:.6f} Accuracy: {accuracy:.2f} {train_acc:.2f} % Best test Accuracy: {best_accuracy:.2f}%')
+        
+        iter_ = 1
+        accuracy = test(model, dataset.test2, criterion)
+        train_acc = test(model, dataset.train, criterion)
+
+        print(f'Train Epoch: {iter_}/{end_iter}  Accuracy: {accuracy:.2f} {train_acc:.2f} % Best test Accuracy: {best_accuracy:.2f}%')
         #writer.add_scalar('Accuracy/test', best_accuracy, comp1)
         bestacc[_ite]=best_accuracy
 
@@ -240,28 +255,29 @@ def main(args):
         best_accuracy = 0
         all_loss = np.zeros(end_iter,float)
         all_accuracy = np.zeros(end_iter,float)
-        
+
         checkdir(f"{os.getcwd()}/dumps/{args.prune_type}/{args.noise_type}/{args.arch_type}/{args.dataset}/{args.kl}+{args.prior}/")
         all_loss.dump(f"{os.getcwd()}/dumps/{args.prune_type}/{args.noise_type}/{args.arch_type}/{args.dataset}/{args.kl}+{args.prior}/{args.prune_type}_all_loss{str(_ite)}.dat")
 
     # Dumping Values for Plotting
     comp.dump(f"{os.getcwd()}/dumps/{args.prune_type}/{args.noise_type}/{args.arch_type}/{args.dataset}/{args.kl}+{args.prior}/{args.prune_type}_compression.dat")
     bestacc.dump(f"{os.getcwd()}/dumps/{args.prune_type}/{args.noise_type}/{args.arch_type}/{args.dataset}/{args.kl}+{args.prior}/{args.prune_type}_bestaccuracy.dat")
-    
+    #ipdb.set_trace()
+    torch.save(mask, f"{os.getcwd()}/saves/{args.arch_type}/{args.dataset}/{args.kl}+{args.prior}/{args.prune_type}/mask.pth.tar")
 
     # Plotting
     a = np.arange(args.prune_iterations)
-    plt.plot(a, bestacc, c="blue", label="Winning tickets") 
+    plt.plot(a, bestacc, c="blue", label="Winning tickets")
     plt.title(f"Test Accuracy vs Unpruned Weights Percentage ({args.dataset},{args.arch_type},{args.prune_type}, Kl:{args.kl}, Prior:{args.prior})")
-    plt.xlabel("Unpruned Weights Percentage") 
-    plt.ylabel("test accuracy") 
-    plt.xticks(a, comp, rotation ="vertical") 
+    plt.xlabel("Unpruned Weights Percentage")
+    plt.ylabel("test accuracy")
+    plt.xticks(a, comp, rotation ="vertical")
     plt.ylim(0,100)
-    plt.legend() 
-    plt.grid(color="gray") 
+    plt.legend()
+    plt.grid(color="gray")
     checkdir(f"{os.getcwd()}/plots/{args.prune_type}/{args.initial}/{args.noise_type}/{args.arch_type}/{args.dataset}/{args.kl}+{args.prior}/acc")
-    plt.savefig(f"{os.getcwd()}/plots/{args.prune_type}/{args.initial}/{args.noise_type}/{args.arch_type}/{args.dataset}/{args.kl}+{args.prior}/acc/{args.prune_type}KL{args.kl}P{args.prior}_AccuracyVsWeights.png", dpi=1200) 
-    plt.close()                    
+    plt.savefig(f"{os.getcwd()}/plots/{args.prune_type}/{args.initial}/{args.noise_type}/{args.arch_type}/{args.dataset}/{args.kl}+{args.prior}/acc/{args.prune_type}KL{args.kl}P{args.prior}_AccuracyVsWeights.png", dpi=1200)
+    plt.close()
 
 
 
@@ -274,14 +290,14 @@ if __name__=="__main__":
     parser.add_argument("--weight_decay", default=1e-4, type=float, help="Weight Decay")
     parser.add_argument("--lr_p", default=1e-2, type=float, help="lr for posterier variance")
     parser.add_argument("--batch_size", default=128, type=int)
-    parser.add_argument("--start_iter", default=0, type=int) 
+    parser.add_argument("--start_iter", default=0, type=int)
     parser.add_argument("--end_iter", default=25, type=int)
     parser.add_argument("--print_freq", default=1, type=int)
     parser.add_argument("--valid_freq", default=1, type=int)
     parser.add_argument("--resume", action="store_true")
     parser.add_argument("--threads", default=4, type=int, help="number of threads of data loader")
     parser.add_argument("--prune_type", default="lt", type=str, help="lt |noise|random")
- 
+
     parser.add_argument("--initial", default="last", type=str, help="reinit|original|last|rewind")
     parser.add_argument("--initial_p", default="last", type=str, help="reinit|last")
 
@@ -305,7 +321,7 @@ if __name__=="__main__":
     args = parser.parse_args()
 
 
-    os.environ["CUDA_DEVICE_ORDER"]="PCI_BUS_ID"   
+    os.environ["CUDA_DEVICE_ORDER"]="PCI_BUS_ID"
     os.environ["CUDA_VISIBLE_DEVICES"]=args.gpu
-    
+
     main(args)
